* Fix demo playing & make CPreGame a bit more tolerating   towards order of network traffic. * Show server name in "Error looking up server from dns." error msg. * Fix gcc warning in Sm3GroundDrawer.cpp
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@1958 37977431-3df6-0310-b722-df95706aa16b

+++ rts/Map/SM3/Sm3GroundDrawer.cpp
@@ -40,7 +40,7 @@ CSm3GroundDrawer::~CSm3GroundDrawer()
 	if (groundShadowVP)
 		glDeleteProgramsARB( 1, &groundShadowVP );
 
-	configHandler.SetInt("SM3TerrainDetail", tr->config.detailMod * 100);
+	configHandler.SetInt("SM3TerrainDetail", int(tr->config.detailMod * 100));
 }
 
 void CSm3GroundDrawer::Update()
